Revert "feat(calculator): project spotlight 추가"
This reverts commit 2e1c31e53cdb64a03521f5e030fc9b4cf424833d.
