+++ app/src/Calculator/CalculatorInput.tsx
@@ -1,15 +1,10 @@
 import { PrimaryMediumText, Writable, Button, CheckBox } from '@shared/ui-kit';
 import styled from '@emotion/styled';
 import { useAtom } from 'jotai';
-import { SubjectListAtom } from '@/Calculator/atoms/SubjectListAtom';
-import { ProjectSpotlightResult } from '@/Calculator/ProjectSpotlightResult';
-import { isProjectSpotlightOpenAtom } from '@/Calculator/atoms/isProjectSpotlightOpenAtom';
-import { toInteger } from 'lodash-es';
-import { useEffect } from 'react';
+import { Subject, SubjectListAtom } from '@/Calculator/atoms/SubjectListAtom';
 
 const CalculatorInput = () => {
   const [subjectList, setSubjectList] = useAtom(SubjectListAtom);
-  const [projectFocus, setProjectFocus] = useAtom(isProjectSpotlightOpenAtom);
 
   const heads = [
     '프로젝트명',
@@ -35,11 +30,6 @@ const CalculatorInput = () => {
     ]);
   };
 
-  const onBlurChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    if (projectFocus !== toInteger(e.target.id)) return;
-    setProjectFocus(-1);
-  };
-
   const onInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     const { value } = e.target;
     const name = e.target.name as keyof typeof subjectList;
@@ -49,22 +39,6 @@ const CalculatorInput = () => {
         return {
           ...subject,
           [name]: value,
-        };
-      }
-      return subject;
-    });
-    setSubjectList(updatedSubjectList);
-  };
-
-  const onNumberInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const { value } = e.target;
-    const name = e.target.name as keyof typeof subjectList;
-    const id = parseInt(e.target.id);
-    const updatedSubjectList = subjectList.map((subject) => {
-      if (subject.id === id) {
-        return {
-          ...subject,
-          [name]: parseInt(value),
         };
       }
       return subject;
@@ -87,10 +61,6 @@ const CalculatorInput = () => {
     setSubjectList(updatedSubjectList);
   };
 
-  useEffect(() => {
-    setProjectFocus(-1);
-  }, [setProjectFocus]);
-
   return (
     <>
       <Table>
@@ -112,11 +82,8 @@ const CalculatorInput = () => {
                     id={id.toString()}
                     name="name"
                     onChange={onInputChange}
-                    onFocus={() => setProjectFocus(id)}
-                    onBlur={onBlurChange}
                     value={name}
                   />
-                  <ProjectSpotlightResult index={id} keyword={name} />
                 </td>
                 <td>{exp}</td>
                 <td>
@@ -124,7 +91,7 @@ const CalculatorInput = () => {
                     type="number"
                     id={id.toString()}
                     name="score"
-                    onChange={onNumberInputChange}
+                    onChange={onInputChange}
                     value={score}
                   />
                 </td>
