+++ app/src/Calculator/Search/atoms/IsProjectSpotLightOpenAtom.ts
